Add tests for bad inputs to JIT engine
This commit adds tests for the type checking capabilities of the JIT engine.

It fixes a bug as well. If the JIT engine creates a callable and the callable is alilve while the JIT engine is deallocated, callinb the callable will cause a segfault. This is because the callable retains the function pointer, not the function itself. When the LLVM lite engine is deallocated (which happens when the JIT engine is deallocated), the function itself is also deallocated. Calling the function pointer dereferences the function pointer, which results in a seg fault. This commit fixes this bug.  The tests in this commit also implicitly test this.