Add tests for partially fixed size tensors and castable but unsafe inputs
We previously didn't have tests for functions that took in args with partially fixed-size tensor types, e.g. tensor<3x?xf32>.

This commit adds one test case for this situation.

This commit also adds a test to ensure that we can't give an input that can't be safely cast to the type needed by the input function. For example, if we have an MLIR function requiring a float32 and we pass in a float64 that's too large to be represented by a float32, we now get an exception raised.